core-dashboard/requirements.txt
# requirements.txt

# Framework
fastapi[all]
uvicorn==0.24.0

# Baza danych i Migracje
SQLAlchemy==2.0.44
psycopg==3.1.18 # Nowy, asynchroniczny sterownik Postgres (lepszy niż psycopg2)
alembic==1.13.1
greenlet==3.2.4

# Konfiguracja
pydantic-settings
email-validator
python-multipart
# Logika biznesowa i API
requests==2.32.5
numpy==1.26.4

# Autoryzacja i Google
cryptography==46.0.3
passlib==1.7.4 # Brakowało tego, potrzebne do haszowania haseł
python-jose==3.3.0 # Brakowało tego, potrzebne do JWT
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
# Obserwowalność (SRE)
starlette-exporter

